refactor(*): set playerError instance on all extensible layers instead of use it as a singleton
the problem of use as a singleton is that if we have more than one player on the page only the last
would trigger events since we were setting the core on player error component when creating a new
one. The problem was fixed creating a new player error instance for each player on the page and
setting this instance on all extensible layers: core, container and playback.